fix: pass format check and address Bugbot truncation issues
Run Prettier on OrganizationMenuButton and ZeroStatePage, extract
truncation measurement into useTruncatedText, and keep relative
positioning when the card description is expanded.
